+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 
 - Added support for AppUpdatePin policy to prevent Firefox from being updated beyond the specified version, see
   [#241](https://github.com/cadeyrn/enterprise-policy-generator/issues/241)
+- Added support for UseSystemPrintDialog policy to use the system print dialog instead of the print preview window, see
+  [#249](https://github.com/cadeyrn/enterprise-policy-generator/issues/249)
 - Added MoreFromMozilla option to UserMessaging policy to show or hide “More from Mozilla” section in the settings, see
   [#248](https://github.com/cadeyrn/enterprise-policy-generator/issues/248)
 - Added temporarily_allow_weak_signatures option in the ExtensionSettings policy to allow installation of extensions